backend/requirements.txt
# ThreatStream Backend Dependencies

# FastAPI & ASGI Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Pydantic for validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Kafka
confluent-kafka==2.3.0

# Google Cloud
google-cloud-aiplatform==1.42.1
google-cloud-firestore==2.14.0
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1

# Monitoring & Logging
structlog==24.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Data processing
numpy==1.26.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
